users/pyproject.toml
[project]
name = "webapp"
version = "0.1.0"
description = ""
authors = [
    {name = "Damian",email = "damiankowalczyk3688@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.1"
gunicorn = "^23.0.0"
python-dotenv = "^1.1.1"
pydantic = "^2.11.7"
email-validator = "^2.2.0"
dependency-injector = "^4.48.1"
mongoengine="^0.29.1"
Flask-Mail="^0.9.1"
pyotp = "^2.9.0"
qrcode = {extras = ["pil"], version = "^7.4.2"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-flask = "^1.3.0"
mypy = "^1.17.1"
testcontainers = "^4.13.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"